Rename the case of Uploader
